requirements.txt
Django
langchain
langchain-community
langchain-ollama
langchain-core
ollama
diffusers
transformers
torch
torchvision
accelerate
Pillow
opencv-python
numpy
requests
python-decouple